requirements.txt
Flask
mysql.connector